+++ lib/stoyco_subscription.dart
@@ -1,4 +1,3 @@
-
 export 'designs/atomic/atoms/inputs/subscription_search_bar.dart';
 export 'designs/atomic/molecules/circular_avatar/subscription_circular_image_with_info.dart';
-export 'pages/subscription_catalog/subscription_catalog_screen.dart';
+export 'pages/subscription_catalog/responsive_subscription_catalog_screen.dart';